Finished on sale section

+++ bmerketo-webshop/Models/ProductModel.cs
@@ -11,7 +11,7 @@ public class ProductModel
     public string? Description { get; set; }
     public decimal? OriginalPrice { get; set; }
     public decimal CurrentPrice { get; set; }
-    public decimal SalePercentage { get; set; }
+    public int SalePercentage { get; set; }
     public string ImageUrl { get; set; } = null!;
     public string CategoryName { get; set; } = null!;
     public ICollection<string> Tags { get; set; } = new HashSet<string>();
